MAINT: Move `linalg` function definitions to `linalg.linalg`
